Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
PyJWT
pydantic-settings
pydantic[email]
python-multipart